[sceptre.zc] In the twirl sequence, rename the AltHold state to Twirling.
If the twirling animation is aborted for some reason (like running out of MP/stamina), it jumps out of the former AltHold state, and eventually ends up in the Ready state again. But, when the alt fire button is pressed again after that, then it jumps immediately to the AltHold state, bypassing the usual animation and not spawning the twirling Sceptre!